% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildConstraintSet.R
\name{buildMomentDerivativeConstFunc}
\alias{buildMomentDerivativeConstFunc}
\title{Constructs constraints functions of (26)}
\usage{
buildMomentDerivativeConstFunc(D, J1, J2)
}
\arguments{
\item{D}{Non-negative integer}

\item{J1}{Vector of non-negative real values}

\item{J2}{Vector of positive integers}
}
\value{
A list containing the constraints functions associated with J2 then J1
}
\description{
Constructs the constraints functions \eqn{G_{j,1}} and  \eqn{G_{j,2}} defined in (26)
}
\examples{
buildMomentDerivativeConstFunc(5, 1:2, 1:3)
buildMomentDerivativeConstFunc(1, 1:2, 1:3)
}
