% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_bic.R
\name{calculate_bic}
\alias{calculate_bic}
\title{Calculate BIC index for LOESS model}
\usage{
calculate_bic(loess_obj)
}
\arguments{
\item{loess_obj}{An object of class "loess"}
}
\value{
A BIC value
}
\description{
Calculate BIC index for LOESS model
}
\examples{

cars.lo <- loess(dist ~ speed, cars)
calculate_bic(cars.lo)

}
