% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateGPR.R
\name{calculate_gpr}
\alias{calculate_gpr}
\title{Compute relevant expression statistics for each gene.}
\usage{
calculate_gpr(counts)
}
\arguments{
\item{counts}{single-cell expression matrix where rows represent genes
and columns represent cells.}
}
\value{
Return a data.table where each row represents a gene,
with four columns: average expression level,
positive ratio, positive cell count, and gene name.
}
\description{
Compute relevant expression statistics for each gene.
}
\examples{
library(GLP)
counts <- matrix(sample(0:10,3000*1000, replace = TRUE), nrow = 3000, ncol = 1000)
rownames(counts) <- paste0("gene",1:3000)
gene.info <- calculate_gpr(counts)

}
