% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{Cluster}
\alias{Cluster}
\title{Cluster based on input HVG}
\usage{
Cluster(data, gene_list, res = 0.8)
}
\arguments{
\item{data}{an seurat object}

\item{gene_list}{input HVG}

\item{res}{Value of the resolution parameter(default is 0.8), bigger value obtain a larger number of clusters}
}
\value{
A seurat object
}
\description{
Cluster based on input HVG
}
