% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOESS_OP.R
\name{get_loess_op}
\alias{get_loess_op}
\title{Compute optimal span width for LOESS regression}
\usage{
get_loess_op(x, y, span_values = seq(from = 0.01, to = 0.1, by = 0.01))
}
\arguments{
\item{x}{a vector with Predictor Variable}

\item{y}{a vector with Response Variable}

\item{span_values}{Default span values}
}
\value{
the optimal span value
}
\description{
Compute optimal span width for LOESS regression
}
