% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateOptimizationParameters.R
\name{getCIMomentAndDerivatives}
\alias{getCIMomentAndDerivatives}
\title{Title}
\usage{
getCIMomentAndDerivatives(sample, a, m = NULL, d = NULL, nboot = 1000,
  alpha = 0.05, method = c("rectangular", "ellipsoid", "both"),
  mc.cores = 1, bootSample = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
a <- 0 ; m <- 0 ; d <- 1 ; nboot <- 100
sample <- rnorm(1E4, 0, 1)
CI <- getCIMomentAndDerivatives(sample, a,m,d,nboot = nboot,mc.cores = 4, method = "both", bootSample = TRUE)

title <- "Ellipsoidal and Rectangular 95\%-CI of the Estimated Parameters"
plot <- plotCI(CI$bootSample, CI)
plot + geom_point(aes(x = dnorm(a), y = 1-pnorm(a)), colour = "blue") +
geom_text(aes(x = 1.009*dnorm(a), y = 1-pnorm(a), label = "True Value"), colour = "blue")+
labs(x = "Derivative of order 1", y = "Moment of order 0") + ggtitle(title)

}

