% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLP.R
\name{glp}
\alias{glp}
\title{GLP selection of high variable genes}
\usage{
glp(df, npc = 3, nfeatures = 1000)
}
\arguments{
\item{df}{a data.table where each row represents a gene,
with four columns: average expression level, positive ratio,
positive cell count, and gene name.}

\item{npc}{Minimum number of cells in which a gene must be expressed.}

\item{nfeatures}{the number of high variable genes.}
}
\value{
A list containing three elements:
1.Genes.summary: The input data.table with summary statistics for each gene.
2.Genes.regression: A data.table with the main results of the robust LOESS regression.
3.HVG: A vector of highly variable genes identified through the analysis."
}
\description{
GLP selection of high variable genes
}
\examples{
library(GLP)
counts <- matrix(sample(0:10,3000*1000, replace = TRUE), nrow = 3000, ncol = 1000)
rownames(counts) <- paste0("gene",1:3000)
gene.info <- calculate_gpr(counts)
result <- glp(gene.info, 3, 1000)

}
