% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationFunctions.R
\name{GLPPhase1}
\alias{GLPPhase1}
\title{Title}
\usage{
GLPPhase1(constFun, constRHS, constDir, paramConsFun = NULL, x = NULL,
  xf = 10000, IterMax = 100, scale = 196)
}
\arguments{
\item{scale}{}
}
\value{
a basic feasible solution
}
\description{
Title
}
\examples{
# Function objective to maximize
objFun = function(x,...)   # H for the optim with convexity constraint
{
  c = paramObjFun$c
  output = I(c <= x)
  return(output)
}

constFun = rep(list(
function(x,...) 1,
function(x,...) x,
function(x,...) x^2
),2)

constDir <- rep(c("<=", ">="), each = 3)

rate <- 1
mu0 <- 1
mu1 <- 1/rate
mu2 <- 2/rate^2

constRHS <- rep(c(mu0,mu1,mu2), 2)
constLambda <- rep(c(0,0,1),2)
objLambda <- 0

paramObjFun <- list(c = qexp(0.9, rate))

xf <- 1e4
initBFS  <-  GLPPhase1(constFun, constRHS,constDir)
}

