% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustLOESS.R
\name{robust_loess}
\alias{robust_loess}
\title{LOESS with Tukey’s biweight robust statistical method}
\usage{
robust_loess(x, y, span = span, iterations = 1, percentage = 0.2)
}
\arguments{
\item{x}{a vector with Predictor Variable}

\item{y}{a vector with Response Variable}

\item{span}{the parameter α which controls the degree of smoothing.}

\item{iterations}{Number of iterations for outlier detection.}

\item{percentage}{the parameter controlling the window size for outlier detection.}
}
\value{
A list containing three elements:
1.weight: gene weight for LOESS regression(outliers are 0).
2.local_mads: gene's local median absolute deviation.
3.model: An final robust LOESS regression model.
}
\description{
LOESS with Tukey’s biweight robust statistical method
}
