% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationFunctions.R
\name{GLPPhase1}
\alias{GLPPhase1}
\title{Title}
\usage{
GLPPhase1(constFun, constRHS, constDir, paramConsFun = NULL, x = NULL,
  xf = 10000, IterMax = 100, scale = 196)
}
\arguments{
\item{scale}{}
}
\value{
a basic feasible solution
}
\description{
Title
}
\examples{

####
#### Finding a basic feasible solution (p,x) such that
####
#### sum(p)  = 1
#### sum(px) = 1
#### sum(px^2) = 2
####
####

# Function objective to maximize
objFun = function(x,...)   # H for the optim with convexity constraint
{
  c = paramObjFun$c
  output = I(c <= x)
  return(output)
}

# Function for the integrals of the constraints inequality
constFun = rep(list(
function(x,...) 1,
function(x,...) x,
function(x,...) x^2
),2)

# Direction of the inequality constraints
constDir <- rep(c("<=", ">="), each = 3)

# Values on the RHS of each inequality
mu0 <- 1
mu1 <- 1
mu2 <- 2

constRHS <- rep(c(mu0,mu1,mu2), 2)

# Lambdas for the objective function and the constraints functions
constLambda <- rep(c(0,0,1),2)
objLambda <- 0

# Get a basic feasible solution
initBFS  <-  GLPPhase1(constFun, constRHS,constDir)

# Check feasibility
with(initBFS, c(sum(p), sum(p*x), sum(p*x^2)))
}

